--- lib/blueprinter/v2/field_serializers/collection.rb
@@ -1,46 +1,0 @@
-# frozen_string_literal: true
-
-module Blueprinter
-  module V2
-    module FieldSerializers
-      # Serializesr for collection fields
-      class Collection < Base
-        def serialize(ctx)
-          # perf boost from "unrolled" built-in hooks
-          value = @conditionals.around_collection_value(ctx) do |ctx|
-            @defaults.around_collection_value(ctx) do |ctx|
-              # perf boost by skipping `around` when no extensions use it
-              if @hook_around_collection_value
-                @hooks.around(:around_collection_value, ctx) do |ctx|
-                  extract ctx
-                end
-              else
-                extract ctx
-              end
-            end
-          end
-          value.nil? ? nil : blueprint_value(value, ctx)
-        end
-
-        private
-
-        def blueprint_value(value, ctx)
-          field_blueprint = ctx.field.blueprint
-          if @instances.blueprint(field_blueprint).is_a? V2::Base
-            parent = Context::Parent.new(ctx.blueprint.class, ctx.field, ctx.object)
-            child_serializer = @instances.serializer(field_blueprint, ctx.options, ctx.store, ctx.depth + 1)
-            child_serializer.collection(value, parent:, depth: ctx.depth + 1)
-          else
-            opts = { v2_instances: @instances, v2_depth: ctx.depth, v2_store: ctx.store }
-            field_blueprint.hashify(value, view_name: :default, local_options: ctx.options.dup.merge(opts))
-          end
-        end
-
-        # We save a lot of time by skipping hooks that aren't used
-        def find_used_hooks!
-          @hook_around_collection_value = @hooks.registered? :around_collection_value
-        end
-      end
-    end
-  end
-end
