+++ spec/v2/instance_cache_spec.rb
@@ -19,28 +19,6 @@ describe Blueprinter::V2::InstanceCache do
     end
   end
 
-  context '#serializer' do
-    it "returns a new instance of Serializer"  do
-      widget_serializer = subject.serializer(widget_blueprint, { foo: true }, store, 1)
-      expect(widget_serializer).to be_a Blueprinter::V2::Serializer
-      expect(widget_serializer.blueprint.class).to be widget_blueprint
-      expect(widget_serializer.options).to eq({ foo: true })
-      expect(widget_serializer.instances).to be subject
-
-      category_serializer = subject.serializer(category_blueprint, { foo: false }, store, 1)
-      expect(category_serializer).to be_a Blueprinter::V2::Serializer
-      expect(category_serializer.blueprint.class).to be category_blueprint
-      expect(category_serializer.options).to eq({ foo: false })
-      expect(category_serializer.instances).to be subject
-    end
-
-    it "returns the same instance of Serializer" do
-      widget_serializer1 = subject.serializer(widget_blueprint, { foo: true }, store, 1)
-      widget_serializer2 = subject.serializer(widget_blueprint, { foo: false }, store, 1)
-      expect(widget_serializer2).to be widget_serializer1
-    end
-  end
-
   context "#extension" do
     let(:extension_a) { Class.new(Blueprinter::Extension) }
     let(:extension_b) { Class.new(Blueprinter::Extension) }
